Added statistics about my mono model. The best thing is that...
the z-coordinate has a far smaller spread (standard deviation) than the other two, which _should_ be the case since I tried to keep the circles at fixed heights.

Another thing to note is that the z-values are _not_ the same for each of the two arms, surprisingly. Yeah, it's a bit of a pain but w/e. Just assume the average of these as where I'll be sending the end-effectors, I think that's fine.